Move parallel_region implementation into a separate file
