+++ c/parallel/thread_scheduler.cc
@@ -48,52 +48,6 @@ void thread_scheduler::execute_in_current_thread() {
 
 
 //------------------------------------------------------------------------------
-// parallel_region
-//------------------------------------------------------------------------------
-
-// Implementation class for `dt::parallel_region()` function.
-class once_scheduler : public thread_scheduler {
-  private:
-    std::vector<cache_aligned<size_t>> done;
-    thread_task* task;
-
-  public:
-    once_scheduler(size_t nthreads, thread_task*);
-    thread_task* get_next_task(size_t thread_index) override;
-};
-
-once_scheduler::once_scheduler(size_t nth, thread_task* task_)
-  : done(nth, 0),
-    task(task_) {}
-
-thread_task* once_scheduler::get_next_task(size_t i) {
-  if (i >= done.size() || done[i].v) {
-    return nullptr;
-  }
-  done[i].v = 1;
-  return task;
-}
-
-
-
-void parallel_region(function<void()> fn) {
-  parallel_region(0, fn);
-}
-
-void parallel_region(size_t nthreads, function<void()> fn) {
-  thread_pool* thpool = thread_pool::get_instance();
-  xassert(thpool->in_master_thread());
-  size_t nthreads0 = thpool->size();
-  if (nthreads > nthreads0 || nthreads == 0) nthreads = nthreads0;
-  simple_task task(fn);
-  once_scheduler sch(nthreads, &task);
-  thpool->execute_job(&sch);
-}
-
-
-
-
-//------------------------------------------------------------------------------
 // parallel_for_static
 //------------------------------------------------------------------------------
